Fix more bugs and restructure project
