+++ python_architecture/mypy.ini
@@ -1,8 +1,7 @@
 [mypy]
 ignore_missing_imports = False
+mypy_path = ./src
+check_untyped_defs = True
 
-[mypy-pytest.*]
-ignore_missing_imports = True
-
-[mypy-sqlalchemy.*]
+[mypy-pytest.*, mypy-sqlalchemy.*]
 ignore_missing_imports = True
