+++ python_architecture/src/allocation/adapters/repository.py
@@ -2,7 +2,7 @@
 
 from typing import List
 import abc
-from domain import model
+from allocation.domain import model
 
 
 class AbstractRepository(abc.ABC):
