+++ python_architecture/src/allocation/config.py
@@ -1,8 +1,12 @@
 import os
 
+
 def get_postgres_uri():
     host = os.environ.get("DB_HOST", "localhost")
-    port = 54321 if host == "localhost" else 5432
+    if host == "localhost":
+        port = 54321
+    else:
+        port = 5432
 
     password = os.environ.get("DB_PASSWORD", "abc123")
     user, db_name = "allocation", "allocation"
@@ -12,6 +16,9 @@ def get_postgres_uri():
 
 def get_api_url():
     host = os.environ.get("API_HOST", "localhost")
-    port = 5005 if host == "localhost" else 80
+    if host == "localhost":
+        port = 5005 
+    else:
+        port = 80
 
     return f"http://{host}:{port}"
