+++ python_architecture/src/allocation/domain/model.py
