+++ python_architecture/src/setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup
+
+setup(
+    name="allocation",
+    version="0.1",
+    packages=["allocation"],
+)
