+++ python_architecture/tests/conftest.py
@@ -9,8 +9,8 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, clear_mappers
 from sqlalchemy.exc import OperationalError
 
-from adapters.orm import metadata, start_mappers
-import config
+from allocation.adapters.orm import metadata, start_mappers
+from allocation import config
 
 
 @pytest.fixture
@@ -71,6 +71,6 @@ def postgress_session(postgres_db):
 @pytest.fixture
 def restart_api():
     """Looks so, extremely hack-ish!"""
-    (Path(__file__).parent / '../entrypoints/flask_app.py').touch()
+    (Path(__file__).parent / '../src/allocation/entrypoints/flask_app.py').touch()
     time.sleep(0.5)
     wait_for_webapp_to_come_up()
