+++ python_architecture/tests/integration/test_repository.py
@@ -1,5 +1,5 @@
-from domain import model
-from adapters import repository
+from allocation.domain import model
+from allocation.adapters import repository
 
 
 def test_repository_can_save_a_batch(session):
